definición del nombre de la aplicación
