--- src/main/java/org/devdom/controller/exceptions/NonexistentEntityException.java
@@ -1,10 +1,0 @@
-package org.devdom.controller.exceptions;
-
-public class NonexistentEntityException extends Exception {
-    public NonexistentEntityException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    public NonexistentEntityException(String message) {
-        super(message);
-    }
-}
