--- src/main/java/org/devdom/controller/exceptions/PreexistingEntityException.java
@@ -1,10 +1,0 @@
-package org.devdom.controller.exceptions;
-
-public class PreexistingEntityException extends Exception {
-    public PreexistingEntityException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    public PreexistingEntityException(String message) {
-        super(message);
-    }
-}
