--- src/main/java/org/devdom/controller/exceptions/RollbackFailureException.java
@@ -1,10 +1,0 @@
-package org.devdom.controller.exceptions;
-
-public class RollbackFailureException extends Exception {
-    public RollbackFailureException(String message, Throwable cause) {
-        super(message, cause);
-    }
-    public RollbackFailureException(String message) {
-        super(message);
-    }
-}
