--- src/main/java/org/devdom/model/dto/MasterCategory.java
@@ -1,55 +1,0 @@
-package org.devdom.model.dto;
-
-import java.io.Serializable;
-import java.util.List;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/**
- *
- * @author Carlos Vásquez Polanco
- */
-
-@Entity
-@XmlRootElement
-public class MasterCategory implements Serializable {
-
-    @Id
-    private final Long id = 1L;
-    
-    @OneToMany
-    private List<Category> category;
-    
-    @OneToOne
-    private Pagination pagination;
-
-    public List<Category> getCategory() {
-        return category;
-    }
-
-    public void setCategory(List<Category> category) {
-        this.category = category;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * @return the pagination
-     */
-    public Pagination getPagination() {
-        return pagination;
-    }
-
-    /**
-     * @param pagination the pagination to set
-     */
-    public void setPagination(Pagination pagination) {
-        this.pagination = pagination;
-    }
-
-}
