+++ src/main/java/org/devdom/skills/controller/exceptions/NonexistentEntityException.java
@@ -1,0 +1,10 @@
+package org.devdom.skills.controller.exceptions;
+
+public class NonexistentEntityException extends Exception {
+    public NonexistentEntityException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    public NonexistentEntityException(String message) {
+        super(message);
+    }
+}
