+++ src/main/java/org/devdom/skills/controller/exceptions/PreexistingEntityException.java
@@ -1,0 +1,10 @@
+package org.devdom.skills.controller.exceptions;
+
+public class PreexistingEntityException extends Exception {
+    public PreexistingEntityException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    public PreexistingEntityException(String message) {
+        super(message);
+    }
+}
