+++ src/main/java/org/devdom/skills/controller/exceptions/RollbackFailureException.java
@@ -1,0 +1,10 @@
+package org.devdom.skills.controller.exceptions;
+
+public class RollbackFailureException extends Exception {
+    public RollbackFailureException(String message, Throwable cause) {
+        super(message, cause);
+    }
+    public RollbackFailureException(String message) {
+        super(message);
+    }
+}
