+++ src/main/java/org/devdom/skills/util/ICategory.java
@@ -1,0 +1,15 @@
+package org.devdom.skills.util;
+
+/**
+ *
+ * @author Carlos Vásquez Polanco
+ */
+public class ICategory {
+    
+    private ICategory(){
+        throw new AssertionError();
+    }
+    
+    //public static final int ROWS_PER_PAGE = 10;
+    
+}
