+++ src/main/java/org/devdom/skills/util/ISkills.java
@@ -1,0 +1,21 @@
+package org.devdom.skills.util;
+
+/**
+ * 
+ * @author Carlos Vásquez Polanco
+ */
+public final class ISkills{
+    
+    private ISkills(){
+        throw new AssertionError();
+    }
+
+    public static final String test = "";
+    
+    public static final String FIND_SKILLS_BY_CATEGORY_ID = "Skills.findSkillsByCategoryId";
+    
+    public static final String FIND_SKILLS_BY_ID = "Skills.findSkillsById";
+    
+    
+    
+}
