--- src/main/java/org/devdom/util/ICategory.java
@@ -1,15 +1,0 @@
-package org.devdom.util;
-
-/**
- *
- * @author Carlos Vásquez Polanco
- */
-public class ICategory {
-    
-    private ICategory(){
-        throw new AssertionError();
-    }
-    
-    //public static final int ROWS_PER_PAGE = 10;
-    
-}
